Rename getCsp.js to getPermissionHeaders.js
