+++ internals/getPermissionHeaders.js
