+++ internals/scripts/print-extension-csp.js
@@ -1,3 +1,3 @@
 // @ts-check
-const { getCsp } = require('../getCsp.js')
+const { getCsp } = require('../getPermissionHeaders.js')
 console.log(getCsp({ isExtension: true }))
